move io test to tests
