+++ tests/snos.rs
@@ -1,15 +1,18 @@
 mod common;
 
 use blockifier::test_utils::DictStateReader;
-use cairo_felt::Felt252;
-use common::{initial_state, prepare_os_test, utils::check_output_vs_python, utils::print_a_hint};
+use cairo_felt::{felt_str, Felt252};
+use common::{
+    initial_state, prepare_os_test, utils::check_output_vs_python, utils::print_a_hint,
+    EXPECTED_PREV_ROOT, EXPECTED_UPDATED_ROOT, TESTING_BLOCK_HASH,
+};
 
 use cairo_vm::cairo_run::{cairo_run, CairoRunConfig};
 use cairo_vm::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::{
     BuiltinHintProcessor, HintFunc,
 };
 
-use snos::{state::SharedState, SnOsRunner};
+use snos::{io::StarknetOsInput, state::SharedState, SnOsRunner};
 
 use starknet_api::block::BlockNumber;
 use starknet_api::core::{ContractAddress, PatriciaKey};
@@ -69,6 +72,22 @@ fn prepared_os_test(mut prepare_os_test: SharedState<DictStateReader>) {
         stark_felt!("4ed2a0d5f47780aee355c14a37ab2ae7dc8fb6f73773563e02fef51b4ec4abe"),
         delegate_root.0
     );
+}
+
+#[rstest]
+fn parse_os_input() {
+    let input = StarknetOsInput::load("tests/common/os_input.json");
+    assert_eq!(felt_str!(TESTING_BLOCK_HASH, 16), input.block_hash);
+    assert_eq!(
+        felt_str!(EXPECTED_PREV_ROOT, 16),
+        input.contract_state_commitment_info.previous_root
+    );
+    assert_eq!(
+        felt_str!(EXPECTED_UPDATED_ROOT, 16),
+        input.contract_state_commitment_info.updated_root
+    );
+    assert!(input.contracts.get(&Felt252::from(0)).is_some());
+    assert!(input.transactions.len() > 0);
 }
 
 #[rstest]
